+++ Cargo.toml
@@ -19,6 +19,7 @@ exclude = [
 
 [dependencies]
 atomic-counter = "1.0"
+bincode = "1.3"
 bytesize = "1.1"
 byte-unit = "4.0"
 chrono = { version = "0.4", features = ["serde", "clock"] }
@@ -27,6 +28,7 @@ console = "0.15"
 crossbeam-utils = "0.8"
 csv = "1.1"
 dashmap = "5.2"
+dirs = "4.0"
 dtparse = "1.2"
 dunce = "1.0"
 fallible-iterator = "0.2"
@@ -45,6 +47,7 @@ rayon = "1.5"
 regex = "1.5"
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+sled = "0.34"
 smallvec = "1.8"
 stfu8 = "0.2"
 structopt = "0.3"
